backend/requirements.txt
 Django==6.0.2
 djangorestframework==3.15.2
 djangorestframework-simplejwt==5.3.1
 django-cors-headers==4.4.0
razorpay==1.4.2
